feat: init with foundry framework

+++ src/OperatorManager.sol
@@ -1,8 +1,9 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.18;
-import './interface/IOrderlyDex.sol';
-import './library/signature.sol';
-import './library/types.sol';
+
+import "./interface/IOrderlyDex.sol";
+import "./library/signature.sol";
+import "./library/types.sol";
 
 /**
  * OperatorManager is responsible for executing cefi tx, only called by operator.
@@ -15,9 +16,9 @@ contract OperatorManager {
 
     // ids
     // futures_upload_batch_id
-    uint public futures_upload_batch_id;
+    uint256 public futures_upload_batch_id;
     // event_upload_batch_id
-    uint public event_upload_batch_id;
+    uint256 public event_upload_batch_id;
 
     // only operator
     modifier only_operator() {
@@ -32,10 +33,10 @@ contract OperatorManager {
     }
 
     // entry point for operator to call this contract
-    function operator_execute_action(
-        Types.OperatorActionData action_data,
-        bytes calldata action
-    ) public only_operator {
+    function operator_execute_action(Types.OperatorActionData action_data, bytes calldata action)
+        public
+        only_operator
+    {
         if (action_data == Types.OperatorActionData.FuturesTradeUpload) {
             // FuturesTradeUpload
             futures_trade_upload_data(abi.decode(action, (Types.FuturesTradeUploadData)));
@@ -45,19 +46,16 @@ contract OperatorManager {
         } else {
             revert("invalid action_data");
         }
-        
     }
 
     // futures trade upload data
-    function futures_trade_upload_data(
-        Types.FuturesTradeUploadData memory data
-    ) internal {
+    function futures_trade_upload_data(Types.FuturesTradeUploadData memory data) internal {
         require(data.batch_id == futures_upload_batch_id, "batch_id not match");
         Types.FuturesTradeUpload[] memory trades = data.trades; // gas saving
         require(trades.length == data.count, "count not match");
         _validate_perp(trades);
         // process each validated perp trades
-        for (uint i = 0; i < data.count; i++) {
+        for (uint256 i = 0; i < data.count; i++) {
             _process_validated_futures(trades[i]);
         }
         // update_futures_upload_batch_id
@@ -66,54 +64,37 @@ contract OperatorManager {
     }
 
     // validate futres trade upload data
-    function _validate_perp(
-        Types.FuturesTradeUpload[] memory trades
-    ) internal pure {
-        for (uint i = 0; i < trades.length; i++) {
+    function _validate_perp(Types.FuturesTradeUpload[] memory trades) internal pure {
+        for (uint256 i = 0; i < trades.length; i++) {
             // first, check signature is valid
             _verify_signature(trades[i]);
             // second, check symbol (and maybe other value) is valid
-            // TODO  
+            // TODO
         }
     }
 
-    function _verify_signature(
-        Types.FuturesTradeUpload memory trade
-    ) internal pure {
+    function _verify_signature(Types.FuturesTradeUpload memory trade) internal pure {
         // TODO ensure the parameters satisfy the real signature
-        bytes32 sig = keccak256(abi.encodePacked(
-            trade.trade_id,
-            trade.symbol,
-            trade.side,
-            trade.trade_qty
-        ));
-    
+        bytes32 sig = keccak256(abi.encodePacked(trade.trade_id, trade.symbol, trade.side, trade.trade_qty));
+
         require(
-            Signature.verify(
-                Signature.getEthSignedMessageHash(sig),
-                trade.signature,
-                trade.account_id
-            ),
+            Signature.verify(Signature.getEthSignedMessageHash(sig), trade.signature, trade.account_id),
             "invalid signature"
         );
     }
 
     // process each validated perp trades
-    function _process_validated_futures(
-        Types.FuturesTradeUpload memory trade
-    ) internal {
-        orderly_dex.update_user_ledger_by_trade_upload(trade); 
+    function _process_validated_futures(Types.FuturesTradeUpload memory trade) internal {
+        orderly_dex.update_user_ledger_by_trade_upload(trade);
     }
 
     // event upload data
-    function event_upload_data(
-        Types.EventUpload memory data
-    ) internal {
+    function event_upload_data(Types.EventUpload memory data) internal {
         require(data.batch_id == event_upload_batch_id, "batch_id not match");
         Types.EventUploadData[] memory events = data.events; // gas saving
         require(events.length == data.count, "count not match");
         // process each event upload
-        for (uint i = 0; i < data.count; i++) {
+        for (uint256 i = 0; i < data.count; i++) {
             _process_event_upload(events[i]);
         }
         // update_event_upload_batch_id
@@ -122,14 +103,12 @@ contract OperatorManager {
     }
 
     // process each event upload
-    function _process_event_upload(
-        Types.EventUploadData memory data
-    ) internal {
-        uint index_withdraw = 0;
-        uint index_settlement = 0;
-        uint index_liquidation = 0;
+    function _process_event_upload(Types.EventUploadData memory data) internal {
+        uint256 index_withdraw = 0;
+        uint256 index_settlement = 0;
+        uint256 index_liquidation = 0;
         // iterate sequence to process each event. The sequence decides the event type.
-        for (uint i = 0; i < data.sequence.length; i++) {
+        for (uint256 i = 0; i < data.sequence.length; i++) {
             if (data.sequence[i] == 0) {
                 // withdraw
                 orderly_dex.execute_withdraw_action(data.withdraws[index_withdraw], data.event_id);
